Correction of all comments
