+++ Countries InfoTests/Tests/CountriesListViewControllerTests.swift
@@ -93,119 +93,6 @@ class CountriesListViewControllerTests: XCTestCase {
         XCTAssertEqual(cell?.countryNameLabel.text , sut.viewModel.countries.first?.name)
     }
     
-    func testApolloManagerNetworkFetch() {
-        let apManager = ApolloManager()
-        
-        let expectation = self.expectation(description: "This worked")
-        
-        apManager.fetch(query: AllCountriesQuery(), type: AllCountries.self) { (countries) in
-            for country in countries.country {
-                print(country.name)
-            }
-            expectation.fulfill()
-        }
-        
-        
-        waitForExpectations(timeout: 5, handler: nil)
-    }
-    
-    func testApolloManagerFetch() {
-        let apManager = ApolloManagerMock()
-        
-        let expectation = self.expectation(description: "This worked")
-        
-        apManager.fetch(query: AllCountriesQuery(), type: AllCountries.self) { (countries) in
-            for country in countries.country {
-                print(country.name)
-            }
-            expectation.fulfill()
-        }
-        waitForExpectations(timeout: 5, handler: nil)
-    }
-    
-    func testImageDataIsNil()  {
-        let expectation = self.expectation(description: "This worked")
-        var image: UIImage? = nil
-        
-        let imageMock = MockImageFactory()
-        stub(imageMock) { (mock) in
-            mock.imageGenerator(code: anyString(), completion: anyClosure()).then { (_, handler) in
-                handler(nil)
-            }
-        }
-        imageMock.imageGenerator(code: "897789") { (data) in
-            XCTAssertNil(data)
-                if let img = UIImage(named: NameConstants.unknownFlagImage) {
-                    image = img
-            }
-            expectation.fulfill()
-        }
-        
-        waitForExpectations(timeout: 10, handler: nil)
-        XCTAssertNotNil(image)
-    }
-    
-    func testImageDataIsNotNil()  {
-        let expectation = self.expectation(description: "This worked")
-        
-        let imageMock = MockImageFactory()
-        stub(imageMock) { (mock) in
-            mock.imageGenerator(code: anyString(), completion: anyClosure()).then { (_, handler) in
-                handler(Data())
-            }
-        }
-        imageMock.imageGenerator(code: "uk") { (data) in
-            expectation.fulfill()
-            XCTAssertNotNil(data)
-        }
-        waitForExpectations(timeout: 10, handler: nil)
-        XCTAssertNotNil(imageMock)
-    }
-    
-    func testImageDataFromNetworkWhenDataIsCorrect()  {
-        
-        let expectation = self.expectation(description: "This worked")
-        sut.viewModel = CountriesListViewModel(apollo: ApolloManager())
-        sut.viewModel.countries.append(Country(name: "", capital: "", population: 0, currencies: [], officialLanguages: [], timezones: [], callingCodes: [], alpha2Code: "br", subregion: nil))
-        let img = ImageFactory()
-        
-        let cell = sut.countriesCollectionView.dataSource?.collectionView(sut.countriesCollectionView, cellForItemAt: IndexPath(item: 0, section: 0)) as? CountryViewCell
-        cell?.factory = img
-        sut.loadViewIfNeeded()
-    
-        img.imageGenerator(code: "br") { (data) in
-            DispatchQueue.main.asyncAfter(deadline: .now() + 1, execute: {
-                XCTAssertNotNil(cell?.flagImageView.image)
-                expectation.fulfill()
-            })
-        }
-
-        waitForExpectations(timeout: 20, handler: nil)
-        XCTAssertNotNil(img)
-    }
-    
-    func testImageDataFromNetworkWhenDataIsNotCorrect()  {
-        let expectation = self.expectation(description: "This worked")
-        sut.viewModel = CountriesListViewModel(apollo: ApolloManager())
-        sut.viewModel.countries.append(Country(name: "", capital: "", population: 0, currencies: [], officialLanguages: [], timezones: [], callingCodes: [], alpha2Code: "br4455", subregion: nil))
-        let img = ImageFactory()
-        
-        let cell = sut.countriesCollectionView.dataSource?.collectionView(sut.countriesCollectionView, cellForItemAt: IndexPath(item: 0, section: 0)) as? CountryViewCell
-        cell?.factory = img
-        sut.loadViewIfNeeded()
-    
-        img.imageGenerator(code: "br") { (data) in
-            DispatchQueue.main.asyncAfter(deadline: .now() + 1, execute: {
-                XCTAssertEqual(cell?.flagImageView.image, UIImage(named: NameConstants.unknownFlagImage))
-                expectation.fulfill()
-            })
-        }
-
-        waitForExpectations(timeout: 20, handler: nil)
-        XCTAssertNotNil(img)
-
-    }
-    
     func testPushToDetailViewController() {
         let mockApolloManager = ApolloManagerMock()
         sut.viewModel  = CountriesListViewModel(apollo: mockApolloManager)
@@ -220,21 +107,5 @@ class CountriesListViewControllerTests: XCTestCase {
         XCTAssertNotNil(detailViewController.nameLabel)
     }
     
-    func testFetchIsCalled() {
-        let mockApolloManager = ApolloManagerMock()
-        let expe = XCTestExpectation(description: "This Work")
-        mockApolloManager.expectation = expe
-        sut.viewModel  = CountriesListViewModel(apollo: mockApolloManager)
-        wait(for: [expe], timeout: 5)
-    }
-    
-    func testFetchCompletionNotNil() {
-        let mockApolloManager = ApolloManagerMock()
-        sut.viewModel  = CountriesListViewModel(apollo: mockApolloManager)
-        
-        mockApolloManager.fetch(query: AllCountriesQuery(), type: AllCountries.self) { (countries) in
-            XCTAssertNotNil(countries)
-        }
-    }
 }
